// src/app/models/clients.ts
export interface Clients{
    _id?:string
    name: string
    status:string
    person: string
    age: number
    createdAt?:string
    labor_Old: string
    seniority_At_Home: number
    own_House: string
    land_Line: string
    phone:string
    voucher_Valid: string
    marital_Status: string
    no_of_Dependents: number
    from: string
    position: string
    trade_Name: string
    updateAt?:string
}